added zip file extraction and loading as numpy

+++ src/b3get/datasets.py
@@ -1,5 +1,9 @@
 import os
+import glob
 import requests
+import zipfile
+import tifffile
+
 from bs4 import BeautifulSoup
 from b3get.utils import *
 from io import BytesIO
@@ -21,7 +25,7 @@ class dataset():
             raise RuntimeError('No dataset can be reached at {}'.format(baseurl))
 
         self.baseurl = baseurl
-        self.datasetid = baseurl.split('/')[-1]
+        self.datasetid = baseurl.rstrip('/').split('/')[-1]
 
     def title(self):
         """ retrieve the title of the dataset """
@@ -29,7 +33,7 @@ class dataset():
         hdoc = BeautifulSoup(r.text, 'html.parser')
         return hdoc.title.string
 
-    def list_images(self):
+    def list_images(self, absolute_url=False):
         """ retrieve the list of images for this dataset """
         values = []
         r = requests.get(self.baseurl)
@@ -38,12 +42,27 @@ class dataset():
         for anc in all_links:
             href = anc.get('href')
             if len(href) > 0 and "zip" in href and "images" in href:
-                values.append(href)
+                url = href if not absolute_url else "/".join([self.baseurl, href])
+                values.append(url)
         return values
 
-    def pull_images(self, rex=""):
+    def list_gt(self, absolute_url=False):
+        """ retrieve the list of images for this dataset """
+        values = []
+        r = requests.get(self.baseurl)
+        hdoc = BeautifulSoup(r.text, 'html.parser')
+        all_links = hdoc.find_all('a')
+        for anc in all_links:
+            href = anc.get('href')
+            if len(href) > 0 and "zip" in href and ("foreground" in href or "labels" in href):
+                url = href if not absolute_url else "/".join([self.baseurl, href])
+                values.append(url)
+        return values
+
+    def pull_files(self, filelist, rex=""):
         """ given a regular expression <rex>, download the files matching it from the dataset site """
-        imgs = filter_files(self.list_images(), rex)
+
+        imgs = filter_files(filelist, rex)
         done = []
         if len(imgs) == 0:
             print("no images found matching {}".format(rex))
@@ -55,18 +74,111 @@ class dataset():
             os.makedirs(dst)
 
         for zurl in imgs:
-            url = "/".join([self.baseurl,zurl])
+            url = "/".join([self.baseurl, zurl])
+            exp_size = size_of_content(url)
+            fname = os.path.split(zurl)[-1]
+            dstf = os.path.join(dst, fname)
+            if os.path.exists(dstf) and os.path.isfile(dstf) and os.stat(dstf).st_size == exp_size:
+                print('{0} already exists in {1} with the correct size {2} kB, skipping it'.format(fname, dst, exp_size/(1024.*1024.)))
+                done.append(dstf)
+                continue
+
             r = requests.get(url)
             assert r.ok, "unable to access URL: {}".format(url)
             pulled_bytes = BytesIO(r.content)
-            dstf = os.path.join(dst, os.path.split(zurl)[-1])
+
             with open(dstf, 'wb') as fo:
                 fo.write(pulled_bytes.read())
-            print("downloaded {0} to {1}".format(url, dstf))
-            done.append(dstf)
+
+            if os.stat(dstf).st_size == exp_size:
+                print("downloaded {0} to {1} ({2} kB)".format(url, dstf, exp_size/(1024.*1024.)))
+                done.append(dstf)
+            else:
+                print("download of {0} to {1} failed ({2} != {3} B)".format(url, dstf, exp_size, os.stat(dstf).st_size ))
 
         return done
 
+    def extract_files(self, filelist, dstdir):
+
+        value = []
+        if not (os.path.exists(dstdir) and os.path.isdir(dstdir)):
+            print('{0} does not exists, will not extract anything to it')
+            return value
+
+        for fn in filelist:
+            before = set(glob.glob(os.path.join(dstdir, "*")))
+            with zipfile.ZipFile(fn, 'r') as zf:
+                print('extracting ', fn)
+                zf.extractall(dstdir)
+                zf.close()
+            after = set(glob.glob(os.path.join(dstdir, "*")))
+            xpaths = after.difference(before)
+            for entry in xpaths:
+                value.extend(glob.glob(os.path.join(entry, "*tif")))
+
+        return value
+
+    def pull_images(self, rex=""):
+        """ given a regular expression <rex>, download the image files matching it from the dataset site """
+        return self.pull_files(self.list_images(), rex)
+
+    def pull_gt(self, rex=""):
+        """ given a regular expression <rex>, download the ground truth files matching it from the dataset site """
+        return self.pull_files(self.list_gt(), rex)
+
+
+    def extract_images(self):
+        """ check tmp_location for downloaded image zip files, if anything is found, extract them """
+
+        tmp = tmp_location()
+        cands = glob.glob(os.path.join(tmp, self.datasetid,"{did}*images*zip".format(did=self.datasetid)))
+        if not cands:
+            return []
+
+        datasetdir = os.path.join(tmp, self.datasetid)
+
+        if not os.path.exists(datasetdir):
+            os.makedirs(datasetdir)
+
+        return self.extract_files(cands, datasetdir)
+
+    def extract_gt(self):
+        """ given a regular expression <rex>, download the ground truth files matching it from the dataset site """
+
+        tmp = tmp_location()
+        fg = os.path.join(tmp, self.datasetid,"{did}*foreground*zip".format(did=self.datasetid))
+        cands = glob.glob(fg)
+        labs = os.path.join(tmp, self.datasetid,"{did}*labels*zip".format(did=self.datasetid))
+        cands.extend(glob.glob(labs))
+        if not cands:
+            return []
+
+        datasetdir = os.path.join(tmp, self.datasetid)
+
+        if not os.path.exists(datasetdir):
+            os.makedirs(datasetdir)
+
+        return self.extract_files(cands, datasetdir)
+
+    def to_numpy(self, folder, glob_stmt="*tif"):
+        """ given a folder, sort the found .tif files and try to open them with tifffile and return a list of numpy arrays """
+        value = []
+        if not os.path.exists(folder):
+            return value
+
+        glob_stmt = os.path.join(folder, glob_stmt)
+        files = glob.glob(glob_stmt)
+        if not files:
+            print('nothing found at', glob_stmt)
+            return value
+
+        for fn in files:
+            try:
+                value.append(tifffile.imread(fn))
+            except Exception as ex:
+                print('unable to open {0} with tifffile due to {1}'.format(fn, ex))
+
+        return value
 
 class ds_006(dataset):
 
@@ -82,7 +194,6 @@ class ds_008(dataset):
 
     def __init__(self, baseurl=None):
         super().__init__(baseurl=ds_008.__baseurl if not baseurl else baseurl)
-
 
 
 class ds_027(dataset):
